Added enum for every Firework.
And I added methods to get the Display name to display it in a list of fireowrks you can select one from to place it, a method to get the JSON value, and a Util method to get an instance of every Item in the enum. I also left comments in the testing section to explain what everything does.